refactor standalone server
